Create persistentplayerbase_23.json
This is my freighter base layout. I need to connect the two teleporters to each other and be able to travel between them in-game.
One short range teleporter is in the bridge of the freighter, and the other is above at the top of the freighter layout above the bridge.